LPS-46899 Move to validator package (no logical changes)
