+++ portal-impl/src/com/liferay/portlet/dynamicdatamapping/validator/DDMFormValuesValidatorImpl.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.dynamicdatamapping.storage;
+package com.liferay.portlet.dynamicdatamapping.validator;
 
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portlet.dynamicdatamapping.StorageException;
@@ -22,6 +22,8 @@ import com.liferay.portlet.dynamicdatamapping.StorageFieldValueException;
 import com.liferay.portlet.dynamicdatamapping.model.DDMForm;
 import com.liferay.portlet.dynamicdatamapping.model.DDMFormField;
 import com.liferay.portlet.dynamicdatamapping.model.Value;
+import com.liferay.portlet.dynamicdatamapping.storage.DDMFormFieldValue;
+import com.liferay.portlet.dynamicdatamapping.storage.DDMFormValues;
 
 import java.util.Collections;
 import java.util.List;
